Add arrow function tests
